feat: incorporated a UI component for the FungibleTokenCreate method within the TokenCreateCustom contract. (#357)
* dapp-update: reorganized sharedMethods and sharedComponents


* dapp-feat: added UI component for FungibleTokenCreate method


* dapp-fix: fixed imports


---------
