+++ system-contract-dapp-playground/src/components/contract-interaction/hts/token-create-custom/methods/index.tsx
@@ -19,11 +19,7 @@
  */
 
 import { Contract } from 'ethers';
-import GrantTokenKYC from './GrantTokenKYC';
-import MintHederaToken from './MintHederaToken';
 import FungibleTokenCreate from './FungibleTokenCreate';
-import AssociateHederaToken from './AssociateHederaToken';
-import NonFungibleTokenCreate from './NonFungibleTokenCreate';
 
 interface PageProps {
   method: string;
@@ -33,13 +29,11 @@ interface PageProps {
 const HederaTokenCreateMethods = ({ baseContract, method }: PageProps) => {
   return (
     <>
-      {method === 'mint' && <MintHederaToken baseContract={baseContract} />}
-      {method === 'grantKYC' && <GrantTokenKYC baseContract={baseContract} />}
-      {method === 'tokenAssociation' && <AssociateHederaToken baseContract={baseContract} />}
+      {method === 'mint' && <>Mint</>}
+      {method === 'grantKYC' && <>KYC</>}
+      {method === 'tokenAssociation' && <>Association</>}
+      {method === 'non-fungibleTokenCreate' && <>non-fungible</>}
       {method === 'fungibleTokenCreate' && <FungibleTokenCreate baseContract={baseContract} />}
-      {method === 'non-fungibleTokenCreate' && (
-        <NonFungibleTokenCreate baseContract={baseContract} />
-      )}
     </>
   );
 };
