+++ system-contract-dapp-playground/src/utils/contract-interactions/HTS/helpers.ts
@@ -19,7 +19,7 @@
  */
 
 import { isAddress } from 'ethers';
-import { KEY_TYPE_MAP, DEFAULT_IHTS_KEY_VALUE } from './constant';
+import { KEY_TYPE_MAP, DEFAULT_IHTS_KEY_VALUE } from './token-create-custom/constant';
 import {
   IHederaTokenServiceKeyType,
   IHederaTokenServiceKeyValueType,
